post

+++ app/Http/Controllers/TasksController.php
@@ -6,18 +6,29 @@ use Illuminate\Http\Request;
 
 use App\Task;    // 追加
 
+use App\User;    // 追加
+
 class TasksController extends Controller
 {
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
+        //タスク一覧を取得
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+        $user = \Auth::user();
         $tasks = Task::all();
-
-        // タスク一覧ビューでそれを表示
+        $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        /* メッセージ一覧ビューでそれを表示
         return view('tasks.index', [
             'tasks' => $tasks,
-        ]);
+        ]);*/
+        // Welcomeビューでそれらを表示
+        return view('tasks.index', $data);
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -34,7 +45,7 @@ class TasksController extends Controller
     // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
-        // バリデーション
+        //バリデーション
         $request->validate([
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:255',
@@ -42,6 +53,7 @@ class TasksController extends Controller
         
         // タスクを作成
         $task = new Task;
+        $task->user_id = auth()->id();  //追加
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
@@ -86,10 +98,12 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを更新
+        if (\Auth::id() === $task->user_id) {
+        $task->user_id = auth()->id();  //追加
         $task->status = $request->status;    // 追加
         $task->content = $request->content;
         $task->save();
-
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -100,8 +114,10 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         // タスクを削除
+        if (\Auth::id() === $task->user_id) {
         $task->delete();
-
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
